// src/types/theme.ts
export type FontSizes = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl'

export type AspectRatios =
  | 'auto'
  | '16/9'
  | '4/3'
  | '4/5'
  | '2/1'
  | '1/1'
  | '3/4'
  | '9/16'
  | '3/2'
  | '2/3'
  | '21/9'
  | '9/21'
  | '18/9'
  | '9/18'

export type GridSizes =
  | '1'
  | '2'
  | '3'
  | '4'
  | '5'
  | '6'
  | '7'
  | '8'
  | '9'
  | '10'
  | '11'
  | '12'

export type HorizontalAlign = 'left' | 'center' | 'right'

export type SectionPadding = 'none' | 'sm' | 'md' | 'lg'

export type SectionTheme = 'white'

export type ButtonTheme = 'default' | 'mono-100' | 'white' | 'black'